Print results in FP

+++ dhry_1.c
@@ -306,11 +306,11 @@ int main (void)
                         / (float) User_Time;
 #endif
     printf ("Microseconds for one run through Dhrystone: ");
-    //printf ("%6.1f \n", Microseconds);
-    printf ("%d \n", (int)Microseconds);
+    printf ("%f \n", Microseconds);
+    //printf ("%d \n", (int)Microseconds);
     printf ("Dhrystones per Second:                      ");
-    //printf ("%6.1f \n", Dhrystones_Per_Second);
-    printf ("%d \n", (int)Dhrystones_Per_Second);
+    printf ("%6.1f \n", Dhrystones_Per_Second);
+    //printf ("%d \n", (int)Dhrystones_Per_Second);
     printf ("\n");
   }
